% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DT_con.R
\name{cal_MCC}
\alias{cal_MCC}
\title{Performance evaluation from other modeling algorithm Result}
\usage{
cal_MCC(pred, label)
}
\arguments{
\item{pred}{Predictions}

\item{label}{Known-endpoint}
}
\value{
result$ACC:   Predicting Accuracy

result$MIS:   MisClassfication Counts

result$MCC:   Matthew's Correlation Coefficients

result$bACC:  balanced Accuracy
}
\description{
Performance evaluation from other modeling algorithm Result
}
