% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DT_con.R
\name{Con_DT}
\alias{Con_DT}
\title{Construct Decision Tree model with pruning}
\usage{
Con_DT(X, Y, min_split = 10, cp = 0.01)
}
\arguments{
\item{X}{dataset}

\item{Y}{data_Labels}

\item{min_split}{minimum number of node in each leaf}

\item{cp}{pre-defined Complexity Parameter (CP) rpart program}
}
\value{
Decision Tree Model with pruning
  Implemented by rpart
}
\description{
Construct Decision Tree model with pruning
}
\seealso{
\code{rpart}
}
