\name{Data_simple}
\docType{data}
\alias{Data_simple}
\title{QSAR dataset with DILI endpoint for demo}
\description{
  This data set gives the DILI endpoint of various compounds (Most or No DILI-concern)
  with QSAR descriptors generated by MOLD2
}
\usage{rivers}
\format{A List containing two vectors:
 X contains 176 observations and 777 variables.
 Y contains DILI endpoints of 176 observations
}
\source{In-house data}
\references{
  Minjun Chen (2011) \emph{FDA-approved drug labeling for the study of drug-induced liver injury}.
  Drug discovery today
}
\keyword{datasets}