% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DT_con.R
\name{DF_acc}
\alias{DF_acc}
\title{Calling Accuracy and Misclassification from Decision Tree Predictions}
\usage{
DF_acc(pred, label)
}
\arguments{
\item{pred}{Predictions}

\item{label}{Known-endpoint}
}
\value{
result$accuracy: Predicting Accuracy

result$MIS:  MisClassfication Counts
}
\description{
Calling Accuracy and Misclassification from Decision Tree Predictions
}

