% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DT_con.R
\name{DF_calp}
\alias{DF_calp}
\title{T-test for feature selection}
\usage{
DF_calp(X, Y)
}
\arguments{
\item{X}{X variable matrix}

\item{Y}{Y label}
}
\description{
T-test for feature selection
}
