% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DF_ConfPlot.R
\name{DF_ConfPlot_accu}
\alias{DF_ConfPlot_accu}
\title{Decision Forest algorithm: confidence level accumulated plot (accumulated version)}
\usage{
DF_ConfPlot_accu(Pred_result, Label, bin = 20, plot = T, smooth = F)
}
\arguments{
\item{Pred_result}{Predictions}

\item{Label}{known label for Test Dataset}

\item{bin}{How many bins occurred in Conf Plot (Default is 20)}

\item{plot}{Draw Plot if True, otherwise output the datamatrix}

\item{smooth}{if TRUE, Fit the performance curve with smooth function (by ggplot2)}
}
\value{
ACC_Conf: return data Matrix ("ConfidenceLevel", "Accuracy", "Matched Samples") for confidence plot (no plot)

ConfPlot: Draw Confidence Plot if True, need install ggplot2
}
\description{
Draw accuracy curve according to the confidence level of predictions
}
