% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DF_dataPre.R
\name{DF_dataPre}
\alias{DF_dataPre}
\title{Decision Forest algorithm: Data pre-processing}
\usage{
DF_dataPre(X, Y = NULL, thres = 0.95)
}
\arguments{
\item{X}{Training Dataset}

\item{thres}{Correlation Coefficient threshold to filter out high correlated features; default is 0.95}
}
\value{
Keep_feat: qualified features in data matrix after filtering
}
\description{
Decision Forest algorithm: Data pre-processing, remove All-Zero columns/features and high correlated features
}
