% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DF_easy.R
\name{DF_easy}
\alias{DF_easy}
\title{Simple pre-defined pipeline for Decision forest}
\usage{
DF_easy(Train_X, Train_Y, Test_X, Test_Y, mode = "default")
}
\arguments{
\item{Train_X}{Training Dataset}

\item{Train_Y}{Training data endpoint}

\item{Test_X}{Testing Dataset}

\item{Test_Y}{Testing data endpoint}

\item{mode}{pre-defined modeling}
}
\value{
data_matrix training and testing result
}
\description{
This is a script of decision forest for easy use t
}
\examples{
  # data(demo_simple)
  X = iris[,1:4]
  Y = iris[,5]
  names(Y)=rownames(X)

  random_seq=sample(nrow(X))
  split_rate=3
  split_sample = suppressWarnings(split(random_seq,1:split_rate))
  Train_X = X[-random_seq[split_sample[[1]]],]
  Train_Y = Y[-random_seq[split_sample[[1]]]]
  Test_X = X[random_seq[split_sample[[1]]],]
  Test_Y = Y[random_seq[split_sample[[1]]]]

  Result = DF_easy(Train_X, Train_Y, Test_X, Test_Y)
}
