% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DT_con.R
\name{DF_perf}
\alias{DF_perf}
\title{performance evaluation between two factors}
\usage{
DF_perf(pred, label)
}
\arguments{
\item{pred}{Predictions}

\item{label}{Known-endpoint}
}
\value{
result$ACC:   Predicting Accuracy

result$MIS:   MisClassfication Counts

result$MCC:   Matthew's Correlation Coefficients

result$bACC:  balanced Accuracy
}
\description{
performance evaluation between two factors
}
