% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DF_CV_plot.R
\name{DF_Trainsummary}
\alias{DF_Trainsummary}
\title{output summary for Dforest test results}
\usage{
DF_Trainsummary(used_model, plot = T)
}
\arguments{
\item{used_model}{Training result}

\item{plot}{if TRUE (default), draw plot}
}
\description{
Draw plot for Dforest test results
}
