% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DForest.R
\name{Dforest}
\alias{Dforest}
\title{Demo script to lean Decision Forest package
  Demo data are located in data/ folder}
\usage{
Dforest()
}
\description{
Demo script to lean Decision Forest package
  Demo data are located in data/ folder
}
\author{
Leihong.Wu
}
