% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFp1_train.R
\name{DFp1_train}
\alias{DFp1_train}
\title{Decision Forest preferred-1 algorithm: Model training}
\usage{
DFp1_train(X, Y, stop_step = 5, Max_tree = 20, Max_feat = 5,
  min_leaf = 10, cp = 0.1, Filter = F, p_val = 0.05, Method = "bACC",
  Quiet = T, Grace_val = 0.05, imp_accu_val = 0.01,
  imp_accu_criteria = F)
}
\arguments{
\item{X}{Training Dataset}

\item{Y}{Training data endpoint}

\item{stop_step}{How many extra step would be processed when performance not improved, 1 means one extra step}

\item{Max_tree}{Maximum tree number in Forest}

\item{Max_feat}{maximum occurrence of features in the forest}

\item{min_leaf}{minimum leaves in tree nodes}

\item{cp}{parameters to pruning decision tree, default is 0.1}

\item{Filter}{doing feature selection before training}

\item{p_val}{P-value threshold measured by t-test used in feature selection, default is 0.05}

\item{Method}{Which is used for evaluating training process. MIS: Misclassification rate; ACC: accuracy}

\item{Quiet}{if TRUE (default), don't show any message during the process}

\item{Grace_val}{Grace Value in evaluation: the next model should have a performance (Accuracy, bACC, MCC) not bad than previous model with threshold}

\item{imp_accu_val}{improvement in evaluation: adding new tree should improve the overall model performance (Accuracy, bACC, MCC) by threshold}
}
\value{
.$accuracy:     Overall training accuracy

.$pred:         Detailed training prediction (fitting)

.$detail:       Detailed usage of Decision tree Features/Models and their performances

.$models:       Constructed (list of) Decision tree models
}
\description{
Decision Forest algorithm: Model training
}
\examples{
  ##data(iris)
  X = iris[,1:4]
  Y = iris[,5]
  names(Y)=rownames(X)
  used_model = DF_train(X,Y,stop_step=4, Method = "MCC")

}
