% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFp2_train.R
\name{DFp2_train}
\alias{DFp2_train}
\title{Decision Forest preferred-2 algorithm: Model training}
\usage{
DFp2_train(X, Y, stop_step = 5, Max_tree = 20, Max_feat = 5, k = 5,
  grace_threshold = 0.01, Filter = F, p_val = 0.05, Method = "bACC",
  Quiet = T)
}
\arguments{
\item{X}{Training Dataset}

\item{Y}{Training data endpoint}

\item{stop_step}{How many extra step would be processed when performance not improved, 1 means one extra step}

\item{Max_tree}{Maximum tree number in Forest}

\item{Max_feat}{maximum occurrence of features in the forest}

\item{Filter}{doing feature selection before training}

\item{p_val}{P-value threshold measured by t-test used in feature selection, default is 0.05}

\item{Method}{Which is used for evaluating training process. MIS: Misclassification rate; ACC: accuracy}

\item{Quiet}{if TRUE (default), don't show any message during the process}

\item{min_leaf}{minimum leaves in tree nodes}

\item{cp}{parameters to pruning decision tree, default is 0.1}

\item{Grace_ACC}{Grace Value in evaluation: the next model should have a performance (Accuracy) not bad than previous model with threshold}

\item{imp_ACC_accu}{improvement in evaluation: adding new tree should improve the overall model performance (accuracy) by threshold}

\item{Grace_bACC}{Grace Value in evaluation: (Balanced Accuracy)}

\item{imp_bACC_accu}{improvement in evaluation: (Balanced Accuracy)}

\item{Grace_MCC}{Grace Value in evaluation: (MCC)}

\item{imp_MCC_accu}{improvement in evaluation: (MCC)}

\item{Grace_MIS}{Grace Value in evaluation: (MIS)}

\item{imp_MIS_accu}{improvement in evaluation: (MIS)}
}
\value{
Result:     Training Model/Performance by KNN
}
\description{
Decision Forest algorithm: Model training
            In Preferred-2 settings, we used KNN instead of Decision Tree
}
\examples{
  ##data(iris)
  X = iris[,1:4]
  Y = iris[,5]
  names(Y)=rownames(X)
  used_model = DF_train(X,Y,stop_step=4, Method = "MCC")

}
