% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DT_con.R
\name{Pred_DT}
\alias{Pred_DT}
\title{Doing Prediction with Decision Tree model}
\source{
rpart
}
\usage{
Pred_DT(model, X)
}
\arguments{
\item{model}{Decision Tree Model}

\item{X}{dataset}
}
\value{
Decision Tree Predictions
  Different endpoints presented in multiple columns
}
\description{
Doing Prediction with Decision Tree model
}
\seealso{
\code{rpart}
}
