% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DF_dataFs.R
\name{DF_dataFs}
\alias{DF_dataFs}
\title{Decision Forest algorithm: Feature Selection in pre-processing}
\usage{
DF_dataFs(X, Y, p_val = 0.05)
}
\arguments{
\item{X}{Training Dataset}

\item{Y}{Training Labels}

\item{p_val}{Correlation Coefficient threshold to filter out high correlated features; default is 0.95}
}
\value{
Keep_feat: qualified features in data matrix after filtering
}
\description{
Decision Forest algorithm: feature selection for two-class predictions,
        kept statistical significant features pass the t-test
}
